Improved statistics api
